all: add CUDA support to work with GPU buffers
New makefile option CUDA_SUPPORT=1 to use the CUDA API for allocation
of GPU memory buffers. Copying to the GPU buffer (after read) and from
the GPU buffer (before write) is then included in the timings to how
this extra copy impacts the results.

Multiple GPUs can be specified to be used either round robin by
different worker threads (i.e. each worker thread uses a different
GPU) or by different service instances (i.e. all workers of one instance
use the same GPU, but different service instances use different GPUs).

+++ source/Statistics.cpp
@@ -284,7 +284,7 @@ void Statistics::printWholeScreenLiveStats()
 	int numFixedHeaderLines = 3; // to calc how many lines are left to show per-worker stats
 	size_t maxNumWorkerLines; // how many lines we have to show per-worker stats
 	size_t numEntriesPerWorker; // total number for this phase
-	size_t numBytesPerWorker; // total number for this phase
+	uint64_t numBytesPerWorker; // total number for this phase
 
 	workerManager.getPhaseNumEntriesAndBytes(numEntriesPerWorker, numBytesPerWorker);
